Yadamu Code Refactor
Migrate YADAMU architecure.
New architecture consists of a core set of functionality that contains the logic for copy ing data between databases or between a database and an export file. The core functonaliy requires an implementation of the DBInterface, DBParser, TableWriter, and StatementGenereator classes be provided for each supported database. Beta 2.0 includes implementations for Oracle, Postgres, MySQL, SQLServer and MariaDB.