Implemnent Yadamu.doImport() using yadamu.pumpDate().
Replace import specific applicaiton logic with pumpData().
Add a FileReader that impliements the minimal set of YadamuDBI operations required to allow import processing to be performed using PumpData().

Replace FileDBI class with FileReader and FileWriter classes.

Add getReader() method to YadamuDBI. The getReader() method returns the object that will generate the JSON objects required by the dbWrtiter class.

The default is to simply return 'this' for database based readers where application logic drives the ordering of the export event stream and applicaiton logic is used to generate the JSON.

For a file based implementation getReader() returns a CLARINET based reader which generates the required objects from a JSON file.

Remove the  'file' parameter from yadamu.doImport.