Enable TABLES, IDENTIFIER_MAPPING_FILE and IDENTIFIER_TRANSFORMATION
Adds support for the following Parameters:

TABLES:  A list of table to be processed. Tables will be processed in the order they are specified

IDENTIFIER_TRANSFORMATION: Supported values are NONE, UPPERCASE, LOWERCASE. If UPPERCASE or LOWERCASE is specified all identifiers will be transformed accordingly when copying data to the target database.

IDENTIFIER_MAPPING_FILE: Specifies the name of a mapping file that will be used to map identifiers in the source to identifiers in the target.

#86
#87